Fix recalculations when deleting a repeat group
Now when a repeat group is deleted and Safe2014DagImpl evaluation
logic is used, triggerables (conditions and calculations) depended on
the all siblings of the deleted repeat group are evaluated.

This fixes #213

Before this change when a repeat group was deleted, Safe2014DagImpl
evaluated triggerables which were depended only on the deleted repeat
group and its children.
Deleting a repeat group in the middle changes the position of all
following repeats. If there were triggerables depended on the following
repeat groups (their position) then those were not evaluatedi.
Therefore values that had been previously evaluated, were wrong.